MaximumEntropyMomentClosures

Documentation for MaximumEntropyMomentClosures.

MaximumEntropyMomentClosures.NewtonParameters

MaximumEntropyMomentClosures.change_basis!

MaximumEntropyMomentClosures.init_lagrange_multipliers!

MaximumEntropyMomentClosures.lagrange_multipliers!

MaximumEntropyMomentClosures.lagrange_multipliers!

MaximumEntropyMomentClosures.normal_lagrange_multipliers!

MaximumEntropyMomentClosures.NewtonParameters — Type

NewtonParameters{FloatType, IntType}

Store parameters of the Newton search algorithm

MaximumEntropyMomentClosures.change_basis! — Method

change_basis!

Determine orthogonal basis that diagonalizes the hessian matrix. Update all input parameters to represent values in the new basis.

MaximumEntropyMomentClosures.init_lagrange_multipliers! — Method

init_lagrange_multipliers!(α_init::AbstractArray{T}, u::AbstractArray{T}) where T

Initialize Lagrange multipliers α_init for a given array of raw moments u.

MaximumEntropyMomentClosures.lagrange_multipliers! — Method

lagrange_multipliers!

Determine lagrange parameters α using a Newton-based search algorithm with an adaptive change of basis.

MaximumEntropyMomentClosures.lagrange_multipliers! — Method

lagrange_multipliers!(α::AbstractArray{T}, u::AbstractArray{T}, p) where T

Determine the Lagrange multipliers for each moment vector stored as a column vector in the abstract array u.

MaximumEntropyMomentClosures.normal_lagrange_multipliers! — Method

normal_lagrange_multipliers!(α::AbstractVector{T}, uc::AbstractVector{T}) where T

Determine the Lagrange multipliers vector α of a normal distribution over $\mathbb{R}$ for a given vector of central moments uc